学習・モデル評価・分類したいデータの分類

+++ Application_for_15/setting.h
@@ -8,7 +8,7 @@ struct learnModel {
 	int learnNum;
 	int featureNum;
 	int teacherNum;
-	double coefficient;
+	float coefficient;
 	float *params;
 };
